requirements/base.txt
Django==3.2.9
pytz==2021.3
djangorestframework==3.12.4
requests==2.26.0
pillow==8.4.0
django-rest-authemail==2.1.0
django-storages[boto3]==1.12.3
firebase-admin==5.0.3 #To be removed once everything has been migrated
